Fix ping (#89)
* Have downnodes report reasons

Have scr_list_down_nodes report reasons when it should not fail anyway.
This may help with debugging bamboo failures.

* Change ping -c 1 to ping -c 1 || ping -c 1

On some networks the first packet of an ICMP request
is dropped due to stale arp information. We've seen this
on omnipath in particular. Can't use ping -c 2 because that
is slow even when it succeeds the first time.

Will continue looking into faster solutions in the slow case.

* add timeouts to ping commands