Add developer documentation
This adds a second set of documentation for developers. It
keeps them in a separate folder to allow for a different
url and thus a different pdf to be generated so that the
developer docs are not distrubuted to users.

Also adds the doc build status badges to the README.