Fix compile error when using BBAPI
This updates some out-of-date scr function names that get called
when using BBAPI, as well as updating the the call to BBAPI's
GetLastErrorDetails function.

Also adds the hints to where the the BBAPI Libraries might be found
to avoid needing to provide the -DWITH_BBAPI_PREFIX option (as was
done in AXL).