add cache bypass support
This is to enable cache-bypass so that an application can write
checkpoint/output datasets directly to the parallel file system and read
them directly from the file system on a restart.

On output:
- route_file returns a PFS path, so the app writes directly to the file
  system
- no redundancy scheme is applied to data files in complete_output,
  though redundancy is applied to metadata and stored in cache
- summary and rank2node metadata files are still created for the dataset
  in the prefix directory as though it had been flushed through normal
  mechanisms, which enables the checkpoint to be fetched into cache on a
  later run if so desired

On restart:
- considers both bypassed and non-bypassed checkpoints
- datasets on PFS are not rebuilt, though associated metadata and
  filemaps are rebuilt
- if restarting from a bypass, route_file returns a PFS path for each
  file, so the app reads files from the file system

On scavenge:
- no data files are copied from compute nodes, though filemaps are
  copied and rebuilt if needed, index file is updated appropriately

The internal cindex structure is updated to mark and query whether a
dataset is stored in cache or if it has bypassed the cache for the
parallel file system.

Various metadata and redundancy encodings of filemaps are still kept in
cache, even for datasets that are bypassed. This design simplifies flush,
rebuild, and scavenge logic, though it means some cache space is still
required.

The bypass mode is now the default, unless the user specifies a
checkpoint descriptor. Using bypass as default enables one to use SCR
out-of-the-box even on machines that have no cache.

A checkpoint descriptor is defined if the user has either defined CKPT
lines in a config file or has defined any of a number of parameters
including SCR_CACHE_BASE, SCR_CACHE_SIZE, SCR_COPY_TYPE, SCR_GROUP,
and SCR_SET_SIZE which are taken to imply the user is trying to use
the cache. In this case, bypass is disabled.

If one has defined checkpoint descriptors and still wants to use bypass,
they can selectively add "BYPASS=1" to the specific checkpoint
descriptor line in a config file meaning bypass should be applied to any
dataset which would select that descriptor. Or one can set
"SCR_CACHE_BYPASS=1" to apply to all checkpoint descriptors (quick way
to turn on bypass without making lots of changes to an existing config
file).

Even in bypass mode, SCR continues to manage cache directories for
bypassed datasets on the compute nodes. It stores metadata in cache in
this case, rather than user data files, so space requirements are
minimal. The default cache directory should be the same as the default
control directory, and both should be ramdisk for performance reasons.
SCR selects a checkpoint descriptor for the bypassed dataset in the
normal way, and that descriptor determines how to encode metadata that
is in cache.