change default cache/control dir from /tmp to /dev/shm
Switch default cache and control directories from /tmp to /dev/shm.
The control directory must be in a node-local location and it should
really be in ramdisk since files are frequently accessed. Also, it's
good to pick a well-known node-local option. Usually /tmp is defined
on all Linux clusters, but it is not always node-local nor is it
always shared memory. The /dev/shm mount makes for a better option at
this point.