add SCR_CURRENT
This defines a new SCR_CURRENT configuration parameter that can be used
to name the checkpoint SCR should restart from. If set, it sets the
"current" maker within the index file under the prefix directory,
overwriting any existing setting that might already be there. This is
most typically used to restart from a checkpoint that is older than the
most recent one that is available.

This provides an alternative to having to execute the
scr_index --current <name> command before starting the job. One can now
set SCR_CURRENT as an environment variable or in a config file and soon
through SCR_Config().

Given that this might be set in a config file or implied in the
application command line throughout the duration of a job, we need to
be careful to only process this once within the job allocation. If a
user names a checkpoint to start the job, that job may write a newer
checkpoint after it has started. If that job then dies and gets
restarted in the same job allocation, we want to restart from the
latest checkpoint, not the old one still named in SCR_CURRENT. To
handle this, we write a marker into the cache index file within the
control directory on each node. We set the marker after processing
SCR_CURRENT, and skip processing SCR_CURRENT in any job where we
detect the marker.

By placing this file in the control directory, rather than the prefix
directory, the marker should be deleted by the system at the end of the
job allocation. A job in a new allocation should once again process any
SCR_CURRENT setting on its first run.

Because this file is in the control directory, we need to restore its
value on all nodes while rebuilding the cache. If we restart using a
spare node, we need that node to get a copy of the marker. We take the
value of the smallest rank that has a value defined.