add function to delete datasets from prefix directory
This adds support for deleting datasets from the prefix directory.

This enables SCR to maintain a sliding window of checkpoints on the
file system in the prefix directory by deleting old checkpoints as new
ones are flushed. After each successful flush, any checkpoint that
falls outside of the window is deleted. Checkpoints that are also
marked as output are not considered for deletion, only pure
checkpoints. One must enable this by specifying the window size with a
new SCR_PREFIX_SIZE parameter. Setting this to 0 disables SCR from
deleting checkpoints in the prefix directory.

The scavenge operation does not currently delete checkpoints, even if
the sliding window is enabled.

It adds a new SCR_Delete(const char* name) API that lets the
application ask SCR to delete a dataset. This deletes user files and
child directories (up to prefix directory) as well as SCR metadata
associated with the dataset, including the ".scr/scr.dataset.id"
directory and the entry from the index file.

It adds a new SCR_Drop(const char* name) API that the application can
use to instruct SCR to drop an entry from the index file for a dataset
while leaving its files in place. This is equivalent to running the
"scr_index --remove name" command.