capture stat in complete, check for changes when deleting files
This records the stat info (uid, gid, mode, atime, mtime, ctime) in
the scr_meta for each file during scr_complete_output().

When deleting a dataset from cache during scr_cache_delete(), we now
check that the mtime and ctime have the same values as recorded in
scr_complete_output(). If not, the user likely changed the file after
declaring it to be complete. That would be a application bug in the
usage of the SCR API, so we print a warning for each such file.