scr: use abtoull for all byte-like parameters
These are right now:

* MPI_BUF_SIZE
* FILE_BUF_SIZE
* FLUSH_ASYNC_BW

for each conversion check that the actual C variable can hold the value
assigned to it.

This changes user visible behaviour for FLUSH_ASYNC_BW which is stored
in a double and before could be set to a fractional bytes/s value. It
will still accept a double but will truncate to an integer value before
setting the double.