Add scr_poststage
This adds in the scr_poststage script and all assocated changes.  This
allows a user to start a checkpoint transfer right before their job
ends and have the transfer continue "in the background" after the job.
The checkpoint files are then finalized using the scr_poststage script
after the transfer ends.  All this uses the IBM Burst Buffer software.

Specific changes:

- Added 'scr_flush_file -r' to resume and finalize a transfer.
- Added 'scr_flush_file -s' to write a summary file for a transfer
  that has been resumed and finalized.
- Added examples/test_scr_poststage.sh for testing scr_poststage.sh
- Added a SCR_FLUSH_POSTSTAGE=1 config option to tell SCR to start
  the transfers but let scr_poststage.sh do the finalizing.
- Update documentation
- Add scr_poststage.sh