initial test with some integration of flux.
Using flux code in the base joblauncher class.
This is to allow pdsh to have varied function with launchers.
Modified scr_hostlist.py to (at least partially) work with numbers only.
The nodetests fails the pings, getting either just a zero or mpi ranks instead of IPs.
If Flux is in use we (must) use Flux to kill/check on the job.
Added a processwait() method in joblauncher base class.