add to flux.py class + script to test launcher.parallel_exec
we can send all rank's stdout/stderr to a file when doing a flux launch
The outputs are mixed among ranks without identification.
put hack script to prepend lines of stdout with the process' rank.
this can emulate a pdsh, dshbak would require condensing.
This should be a feature of Flux and some of these changes can go away