Treat '=' character as legal value in SCR_Config
Previously, the parser in SCR_Config was treating '=' as an invalid
character within a value for a given token.  This caused calling
applications to fail when they passed in directory path names that
contained an '=' character.

This failure behavior is inconsistent because when the application uses
environment variables instead of SCR_Config, this particular character
is allowed.

The fix was to remove the check when the parser is either in 'before_value' or
'in_value' state.

Add tests for SCR_Config with '=' character

Update SCR_Config docs describing '=' char as being valid