slurm.py: Handle case where no end time exists
The allocation may have no time limit, in which case the output by sinfo
is the string "Unknown" which we should not attempt to translate to
seconds.

For the longer term, we may be able to get this information from a
library and avoid parsing altogether as well as the fork/exec, so I left
a comment as a reminder.