scr_common.py: Add choose_bindir() to use build binaries
The python scripts must be able to run binaries from the build directory
(under CMAKE_BINARY_DIR) when "make check" or "make test" are run,
rather than assuming they are in the install directory as they do now.

However when running from an installed instance, they must use
the usual X_BINDIR path.

Add a function to determine which is the case and return an appropriate
bindir.