scr_run.py: exit if no jobid found
We need to log the jobid for later troubleshooting, and need to be
running within an allocation for operations such as scavenge.

If we cannot identify the job id, at least one of those needs is not
being met, so we should exit with an error.

Also, when scr_run.py is unable to determine the nodeset, include the
jobid in the error message as that may help debug the issue.