scr_run.py: exit with non-zero exit code on failure
For use in testing, scr_run.py should exit with a nonzero
exit code if the launched job fails, so ctest can detect
that and report it.

Similarly, users running scru_run.py from a script should
be able to detect failure so the script can do whatever is
appropriate.