Update jobid and endtime based on flux v0.46.1
As of flux-core v0.45.0, flux stopped setting FLUX_JOB_ID when in an
allocation but not in a job.  See flux-core commit

* 485743d19 broker: clear job environment vars in rc scripts

For example,

$ flux mini alloc -n1 -t3h bash -c set | grep FLUX

would no longer include FLUX_JOB_ID.

In addition, the "R" obtained with the "resource.status" RPC stopped
including the expiration time of the allocation, although I'm not sure
as of what flux version or commit.

Updated to interfaces that work properly with flux-core v0.46.1.