Docs: Update to use yaml files
Read the Docs has moved to a yaml file format that is needed to
continue building our docs in the future.

- Adds yaml files for doc and doc-dev
- Fix reference link
- Remove static path from developer docs
- Update developer docs version

NOTE: fail_on_warning and builds for pull requests can be enabled in
the future if desired. This will error on doc changes that violate
rst syntax (can be finicky) and add a build status report to open
pull requests.