tests: fix argument passing to scr_run.py
Problem: the scr_run_py.sh script sometimes passes an extra argument
to scr_run.py. The script passes \"$@\", which comes out to the
string "" when $@ is empty. The string is forwarded on and eventually
causes an error in 'test_api_multiple.c'.

Solution: when $@ is empty, don't pass anything to scr_run.py.